backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-cloud-spanner==3.33.0
gunicorn==21.2.0